+++ CHANGELOG.md
@@ -8,4 +8,5 @@
 - Refactor the wasmbinding implementation to allow multiple msg and query types
 - Add EVM wasmbinding queries
 - Add bech32 wasmbinding queries
-- Added IBC precompile to transfer via EVM
+- Add IBC precompile to transfer via EVM
+- Add bech32 wasmbinding queries